LayoutTests/imported/w3c/IndexedDB-private-browsing/idbcursor_delete_objectstore5-expected.txt


PASS IDBCursor.delete() - object store - throw InvalidStateError when the cursor is being iterated
